Extract of CachingDomainList
